Refactor add/remove methods

+++ src/main/java/com/epam/tishkin/server/service/impl/AuthorServiceImpl.java
@@ -4,7 +4,10 @@ import com.epam.tishkin.models.Author;
 import com.epam.tishkin.server.repository.AuthorRepository;
 import com.epam.tishkin.server.service.AuthorService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 
+import javax.persistence.EntityExistsException;
+import javax.persistence.EntityNotFoundException;
 import java.util.Optional;
 
 public class AuthorServiceImpl implements AuthorService {
@@ -16,6 +19,17 @@ public class AuthorServiceImpl implements AuthorService {
     }
 
     @Override
+    public ResponseEntity<String> addAuthor(Author author) {
+        Optional<Author> currentAuthor = authorRepository.findById(author.getName());
+        if (currentAuthor.isPresent()) {
+            throw new EntityExistsException("Author already exists: " + author.getName());
+        }
+        authorRepository.save(author);
+        return ResponseEntity.ok("Author was added: " + author.getName());
+    }
+
+    /*
+    @Override
     public Author addAuthor(Author author) {
         Optional<Author> currentAuthor = authorRepository.findById(author.getName());
         if (currentAuthor.isPresent()) {
@@ -23,9 +37,14 @@ public class AuthorServiceImpl implements AuthorService {
         }
         return null;
     }
+     */
 
     @Override
-    public void deleteAuthor(String authorName) {
-        authorRepository.delete(new Author(authorName));
+    public ResponseEntity<String> deleteAuthor(String authorName) {
+        if (authorRepository.findById(authorName).isPresent()) {
+            authorRepository.delete(new Author(authorName));
+            return ResponseEntity.ok("Author was deleted: " + authorName);
+        }
+        throw new EntityNotFoundException("Not found: " + authorName);
     }
 }
